etl/requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
importlib-metadata==3.1.1
kombu==4.6.11
mysql-connector-python==8.0.22
numpy==1.19.4
pandas==1.1.5
protobuf==3.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
urllib3==1.26.2
vine==1.3.0
zipp==3.4.0
